requirements.txt
aerich==0.6.3
aiosqlite==0.17.0
anyio==3.6.1
asyncpg==0.26.0
autopep8==1.7.0
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.1.0
click==8.1.3
dictdiffer==0.9.0
distlib==0.3.5
ecdsa==0.18.0
fastapi==0.79.0
filelock==3.8.0
h11==0.13.0
identify==2.5.3
idna==3.3
iso8601==1.0.2
itsdangerous==2.1.2
pip==22.1.2
platformdirs==2.5.2
pyasn1==0.4.8
pycodestyle==2.9.1
pydantic==1.9.2
pypika-tortoise==0.1.6
python-dotenv==0.20.0
python-jose==3.3.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
rsa==4.9
sentry-sdk==1.9.5
setuptools==61.2.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
tomlkit==0.11.4
tortoise-orm==0.19.2
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
virtualenv==20.16.3
wheel==0.37.1
